cursor-3d-animation/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# AI and LLM
openai==1.3.5
python-dotenv==1.0.0

# Video processing
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# Task queue and background jobs
celery==5.3.4
redis==5.0.1

# File handling
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code execution
docker==6.1.3

# Authentication and Database
supabase==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
uuid==1.30